Renamed some stuff and purged old classes
template_rules renamed to rules
